+++ ChangeLog.md
@@ -4,7 +4,12 @@ Follow the principles in [keepachangelog.com](https://keepachangelog.com)!
 
 # v.Next (Current)
 
+### Fixed
+ - Fixed support for the `target` HTML arg on `NavLink` and it's derivatives.
+
 # [v0.2.0-beta.65](https://github.com/moxb/moxb/releases/tag/v0.2.0-beta.65) (2021-02-10)
+
+### Added
 - OneOfFormAnt: add support for vertical layout of choices (like OneOfAnt)
 
 # [v0.2.0-beta.64](https://github.com/moxb/moxb/releases/tag/v0.2.0-beta.64) (2021-02-05)