+++ ChangeLog.md
@@ -14,6 +14,9 @@ Follow the principles in [keepachangelog.com](https://keepachangelog.com)!
  - Fixed support for the `target` HTML arg on `NavLinkButtonAnt` and it's derivatives.
  - Fixed the disabled property on Anchor
 
+### Added
+ - Add some utility functions for dealing with decisions
+
 # [v0.2.0-beta.65](https://github.com/moxb/moxb/releases/tag/v0.2.0-beta.65) (2021-02-10)
 
 ### Added