+++ ChangeLog.md
@@ -4,6 +4,8 @@ Follow the principles in [keepachangelog.com](https://keepachangelog.com)!
 
 # v.Next (Current)
 
+# [v0.2.0-beta.66](https://github.com/moxb/moxb/releases/tag/v0.2.0-beta.66) (2021-02-16)
+
 ### Added
  - BoundLink: added help, enabled and disabled options.
  - OneOfAnt and OneOfButtonAnt: add support for adding a line break after every nth option.